+++ service/mocks/node_forward_mock.go
@@ -5,37 +5,38 @@
 package mocks
 
 import (
+	context "context"
+	reflect "reflect"
+
 	v1 "code.vegaprotocol.io/protos/vega/api/v1"
 	v10 "code.vegaprotocol.io/protos/vega/commands/v1"
-	context "context"
 	gomock "github.com/golang/mock/gomock"
-	reflect "reflect"
 )
 
-// MockNodeForward is a mock of NodeForward interface
+// MockNodeForward is a mock of NodeForward interface.
 type MockNodeForward struct {
 	ctrl     *gomock.Controller
 	recorder *MockNodeForwardMockRecorder
 }
 
-// MockNodeForwardMockRecorder is the mock recorder for MockNodeForward
+// MockNodeForwardMockRecorder is the mock recorder for MockNodeForward.
 type MockNodeForwardMockRecorder struct {
 	mock *MockNodeForward
 }
 
-// NewMockNodeForward creates a new mock instance
+// NewMockNodeForward creates a new mock instance.
 func NewMockNodeForward(ctrl *gomock.Controller) *MockNodeForward {
 	mock := &MockNodeForward{ctrl: ctrl}
 	mock.recorder = &MockNodeForwardMockRecorder{mock}
 	return mock
 }
 
-// EXPECT returns an object that allows the caller to indicate expected use
+// EXPECT returns an object that allows the caller to indicate expected use.
 func (m *MockNodeForward) EXPECT() *MockNodeForwardMockRecorder {
 	return m.recorder
 }
 
-// HealthCheck mocks base method
+// HealthCheck mocks base method.
 func (m *MockNodeForward) HealthCheck(arg0 context.Context) error {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "HealthCheck", arg0)
@@ -43,38 +44,39 @@ func (m *MockNodeForward) HealthCheck(arg0 context.Context) error {
 	return ret0
 }
 
-// HealthCheck indicates an expected call of HealthCheck
+// HealthCheck indicates an expected call of HealthCheck.
 func (mr *MockNodeForwardMockRecorder) HealthCheck(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "HealthCheck", reflect.TypeOf((*MockNodeForward)(nil).HealthCheck), arg0)
 }
 
-// LastBlockHeight mocks base method
-func (m *MockNodeForward) LastBlockHeight(arg0 context.Context) (uint64, error) {
+// LastBlockHeight mocks base method.
+func (m *MockNodeForward) LastBlockHeight(arg0 context.Context) (uint64, int, error) {
 	m.ctrl.T.Helper()
 	ret := m.ctrl.Call(m, "LastBlockHeight", arg0)
 	ret0, _ := ret[0].(uint64)
-	ret1, _ := ret[1].(error)
-	return ret0, ret1
+	ret1, _ := ret[1].(int)
+	ret2, _ := ret[2].(error)
+	return ret0, ret1, ret2
 }
 
-// LastBlockHeight indicates an expected call of LastBlockHeight
+// LastBlockHeight indicates an expected call of LastBlockHeight.
 func (mr *MockNodeForwardMockRecorder) LastBlockHeight(arg0 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
 	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "LastBlockHeight", reflect.TypeOf((*MockNodeForward)(nil).LastBlockHeight), arg0)
 }
 
-// SendTx mocks base method
-func (m *MockNodeForward) SendTx(arg0 context.Context, arg1 *v10.Transaction, arg2 v1.SubmitTransactionRequest_Type) (string, error) {
+// SendTx mocks base method.
+func (m *MockNodeForward) SendTx(arg0 context.Context, arg1 *v10.Transaction, arg2 v1.SubmitTransactionRequest_Type, arg3 int) (string, error) {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "SendTx", arg0, arg1, arg2)
+	ret := m.ctrl.Call(m, "SendTx", arg0, arg1, arg2, arg3)
 	ret0, _ := ret[0].(string)
 	ret1, _ := ret[1].(error)
 	return ret0, ret1
 }
 
-// SendTx indicates an expected call of SendTx
-func (mr *MockNodeForwardMockRecorder) SendTx(arg0, arg1, arg2 interface{}) *gomock.Call {
+// SendTx indicates an expected call of SendTx.
+func (mr *MockNodeForwardMockRecorder) SendTx(arg0, arg1, arg2, arg3 interface{}) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendTx", reflect.TypeOf((*MockNodeForward)(nil).SendTx), arg0, arg1, arg2)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "SendTx", reflect.TypeOf((*MockNodeForward)(nil).SendTx), arg0, arg1, arg2, arg3)
 }